So, here's how it is:
All tests pass.
Layering is kinda broken because our removeBreakLayer and removeThrowLayer seem to not be called as often as they need to be (mostly on the removeThrowLayer side, likely because of implementation for deleting only the first variable, when try--state could be the second variable).
Does not run finally when using break or continue inside a try/catch.

That be how it is.